+++ lib/async_button.dart
@@ -1,13 +1,22 @@
 library async_button;
 
-// Complimentary
-export 'src/helpers/constants.dart';
+// The following (and it's associated) files will be removed in the future versions
+// as the exported classes, functions, etc were deprecated after v1.0.1-beta.2
 export 'src/helpers/async_button_state_style.dart';
-
-// Convenience
 export 'src/core/async_button_state_controller_abstract.dart';
-
-// Necessary
 export 'src/async_elevated_button.dart';
 export 'src/async_outlined_button.dart';
 export 'src/async_text_button.dart';
+// ----
+
+// Complimentary
+export 'src/helpers/constants.dart';
+export 'src/helpers/async_btn_state_style.dart';
+
+// Convenience
+export 'src/core/async_btn_states_controller.dart';
+
+// Necessary
+export 'src/async_elevated_btn.dart';
+export 'src/async_outlined_btn.dart';
+export 'src/async_text_btn.dart';