+++ lib/src/async_outlined_btn.dart
@@ -1,17 +1,13 @@
 import 'package:flutter/material.dart';
 
-import 'core/async_button_core_abstract.dart';
-import 'core/async_button_state_controller_abstract.dart';
-import 'helpers/async_button_state_style.dart';
+import 'core/async_btn_core_abstract.dart';
+import 'core/async_btn_states_controller.dart';
+import 'helpers/async_btn_state_style.dart';
 
-/// Implements [TextButton] for asynchronous [onPressed]
-@Deprecated(
-  'Use AsyncTextBtn instead. '
-  'This feature was deprecated after v1.0.1-beta.2',
-)
-class AsyncTextButton extends AsyncButtonCore {
-  /// Creates an [AsyncTextButton]
-  const AsyncTextButton({
+/// Implements [OutlinedButton] for asynchronous [onPressed]
+class AsyncOutlinedBtn extends AsyncBtnCore {
+  /// Creates an [AsyncOutlinedBtn]
+  const AsyncOutlinedBtn({
     super.key,
     super.loadingStyle,
     super.loadingStyleBuilder,
@@ -34,16 +30,19 @@ class AsyncTextButton extends AsyncButtonCore {
     super.switchOutAnimationDuration,
     super.layoutBuilder,
     super.transitionBuilder,
+    super.onLongPress,
+    super.materialStatesController,
+    super.styleBuilder,
+    super.asyncBtnStatesController,
   });
 
   /// Same as [AsyncOutlinedButton()] but with sample values for [loadingStyle],
   /// [successStyle] and [failureStyle].
-  factory AsyncTextButton.withDefaultStyles({
+  factory AsyncOutlinedBtn.withDefaultStyles({
     Key? key,
-    required Future<void> Function(
-            AsyncButtonStateController btnStateController)?
-        onPressed,
+    required Future<void> Function()? onPressed,
     required Widget child,
+    Future<void> Function()? onLongPress,
     bool switchBackAfterCompletion = true,
     Duration switchBackDelay = const Duration(seconds: 2),
     bool lockWhileAlreadyExecuting = true,
@@ -56,14 +55,16 @@ class AsyncTextButton extends AsyncButtonCore {
     Curve switchOutAnimationCurve = Curves.ease,
     ButtonStyle? style,
     Widget Function(Widget? currentChild, List<Widget> previousChildren)
-        layoutBuilder = AsyncButtonCore.defaultLayoutBuilder,
+        layoutBuilder = AsyncBtnCore.defaultLayoutBuilder,
     Widget Function(Widget child, Animation<double> animation)
-        transitionBuilder = AsyncButtonCore.defaultTransitionBuilder,
+        transitionBuilder = AsyncBtnCore.defaultTransitionBuilder,
+    MaterialStatesController? materialStatesController,
+    AsyncBtnStatesController? asyncBtnStatesController,
   }) {
-    return AsyncTextButton(
+    return AsyncOutlinedBtn(
       onPressed: onPressed,
-      failureStyle: AsyncButtonStateStyle(
-        style: TextButton.styleFrom(
+      failureStyle: AsyncBtnStateStyle(
+        style: OutlinedButton.styleFrom(
           foregroundColor: Colors.red,
         ),
         widget: Row(
@@ -75,8 +76,8 @@ class AsyncTextButton extends AsyncButtonCore {
           ],
         ),
       ),
-      loadingStyle: AsyncButtonStateStyle(
-        style: TextButton.styleFrom(
+      loadingStyle: AsyncBtnStateStyle(
+        style: OutlinedButton.styleFrom(
           foregroundColor: Colors.amber,
         ),
         widget: const SizedBox.square(
@@ -86,8 +87,8 @@ class AsyncTextButton extends AsyncButtonCore {
           ),
         ),
       ),
-      successStyle: AsyncButtonStateStyle(
-        style: TextButton.styleFrom(
+      successStyle: AsyncBtnStateStyle(
+        style: OutlinedButton.styleFrom(
           foregroundColor: Colors.green,
         ),
         widget: Row(
@@ -113,16 +114,19 @@ class AsyncTextButton extends AsyncButtonCore {
       switchOutAnimationCurve: switchOutAnimationCurve,
       transitionBuilder: transitionBuilder,
       layoutBuilder: layoutBuilder,
+      onLongPress: onLongPress,
+      materialStatesController: materialStatesController,
+      asyncBtnStatesController: asyncBtnStatesController,
       child: child,
     );
   }
 
   @override
-  State<AsyncButtonCore> createState() => _AsyncTextButtonState();
+  State<AsyncBtnCore> createState() => _AsyncOutlinedButtonState();
 }
 
-class _AsyncTextButtonState extends AsyncButtonCoreState {
-  static const _simpleButton = TextButton(
+class _AsyncOutlinedButtonState extends AsyncBtnCoreState {
+  static const _simpleButton = OutlinedButton(
     onPressed: null,
     child: SizedBox(),
   );
@@ -130,14 +134,17 @@ class _AsyncTextButtonState extends AsyncButtonCoreState {
   @override
   ButtonStyle intializeButtonStyle() {
     return (widget.style ?? const ButtonStyle())
+        .merge(_simpleButton.themeStyleOf(context))
         .merge(_simpleButton.defaultStyleOf(context));
   }
 
   @override
   ButtonStyleButton makeButton(Widget child) {
-    return TextButton(
+    return OutlinedButton(
       onPressed: onPressed(),
       style: makeButtonStyle(),
+      onLongPress: onLongPressed(),
+      statesController: statesController,
       child: child,
     );
   }